Drop a stray root file and pin zero-size glob paging (#233)
Removes an empty GrepSearch.cs committed at the repository root by
accident, and covers the glob case the issue named but no test pinned:
a page size of zero is a page of one, so page 2 is the second row
rather than the first again. The grep offset widens before it subtracts,
as the other offsets do, and the read_file window helper is named for
what it returns.

+++ CodeExplorer/Search/GrepSearch.cs
@@ -647,6 +647,6 @@ public sealed partial class GrepSearch(IndexReaders readers)
         ///     The files ahead of this page. A <c>long</c>, because <see cref="Page" /> is clamped only from
         ///     below and a large one times the page size wraps an <c>int</c> negative (#233).
         /// </summary>
-        public long Skip => (long)(Page - 1) * PageSize;
+        public long Skip => (Page - 1L) * PageSize;
     }
 }
